requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.38.0
pytest>=8.0.0
requests>=2.31.0
flake8>=7.0.0
